implement DateConverter-Class
